Improvements to Load / Restore
Made general improvements, should for the most part get most scenarios,
though little testing at this point has been done. I was able with this
version to backup user, load user, but may have to reboot program
between procedures, have to work on cleaning out the INI dictionary and
being able to restore and backup on the same computer. Though for usage,
this is not necessary and shouldnt be an issue necessarily